Update dsm.php
Basically, removed most of node, template and plan code that is no
longer relevant to the module. Some of the code still remains because
I've yet to figure out how to safely remove them or change them.

original.php is the original dsm.php prior to removal of the node,
template and plan codes. Do NOT use this file!